requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
joblib==1.4.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
plotly==5.24.1
matplotlib==3.9.2
requests==2.32.3
xgboost==2.1.3
